requirements.txt
glcontext==2.3.7
lxml==4.9.2
moderngl==5.7.4
moderngl-window==2.4.2
multipledispatch==0.6.0
numpy==1.24.1
Pillow==9.4.0
prodict~=0.8.18
pygame==2.1.2
pyglet==2.0.3
PyGLM==2.6.0
PyOpenGL==3.1.6
pyopengltk==0.0.4
pyrr==0.10.3
PyWavefront==1.3.3
six==1.16.0
tk==0.1.0